Bug 560393, part 2: (xpcom directory) Fix compile warning: format $f expects type $t1, but argument 2 has type $t2. r=bsmedberg

+++ xpcom/tests/TestCOMPtr.cpp
@@ -368,7 +368,7 @@ main()
   {
     printf(">>main()\n");
 
-		printf("sizeof(nsCOMPtr<IFoo>) --> %d\n", sizeof(nsCOMPtr<IFoo>));
+		printf("sizeof(nsCOMPtr<IFoo>) --> %u\n", unsigned(sizeof(nsCOMPtr<IFoo>)));
 
 		TestBloat_Raw_Unsafe();
 		TestBloat_Smart();
